+++ R/addpoly.predict.rma.r
@@ -1,15 +1,15 @@
-addpoly.rma         <- function(x,
-row=-2,  level=x$level, annotate, addpred=FALSE, digits, width, mlab,
+addpoly.predict.rma <- function(x,
+rows=-2,                annotate, addpred=FALSE, digits, width, mlab,
 transf, atransf, targs, efac, col, border, lty, fonts, cex, annosym, ...) {
 
    #########################################################################
 
    mstyle <- .get.mstyle("crayon" %in% .packages())
 
-   .chkclass(class(x), must="rma")
+   .chkclass(class(x), must="predict.rma")
 
-   if (!x$int.only)
-      stop(mstyle$stop("Fitted model should not contain moderators."))
+   if (x$pred.type == "scale")
+      stop(mstyle$stop("Cannot add polygons based on predicted scale values."))
 
    if (missing(annotate))
       annotate <- .getfromenv("forest", "annotate", default=TRUE)
@@ -53,41 +53,18 @@ transf, atransf, targs, efac, col, border, lty, fonts, cex, annosym, ...) {
    if (missing(annosym))
       annosym <- .getfromenv("forest", "annosym", default=NULL)
 
-   ddd <- list(...)
-
-   if (!is.null(ddd$addcred))
-      addpred <- ddd$addcred
-
-   if (is.null(ddd$pi.type)) {
-      pi.type <- "default"
-   } else {
-      pi.type <- ddd$pi.type
-   }
-
-   pred <- predict(x, level=level, pi.type=pi.type)
-
-   ci.lb <- pred$ci.lb
-   ci.ub <- pred$ci.ub
-
    if (addpred) {
-      pi.lb <- pred$pi.lb
-      pi.ub <- pred$pi.ub
+      pi.lb <- x$pi.lb
+      pi.ub <- x$pi.ub
    } else {
-      pi.lb <- NA
-      pi.ub <- NA
+      pi.lb <- rep(NA, length(x$pred))
+      pi.ub <- rep(NA, length(x$pred))
    }
 
    #########################################################################
 
-   ### label for model estimate (if not specified)
-
-   if (is.null(mlab))
-      mlab <- sapply(x$method, switch, "FE"="FE Model", "EE"="EE Model", "CE"="CE Model", "RE Model", USE.NAMES=FALSE)
-
-   ### passing ci.lb and ci.ub, so that the bounds are correct when the model was fitted with test="knha"
-
-   addpoly(x$beta, ci.lb=ci.lb, ci.ub=ci.ub, pi.lb=pi.lb, pi.ub=pi.ub,
-           rows=row, level=level, annotate=annotate, digits=digits, width=width,
+   addpoly(x$pred, ci.lb=x$ci.lb, ci.ub=x$ci.ub, pi.lb=pi.lb, pi.ub=pi.ub,
+           rows=rows,             annotate=annotate, digits=digits, width=width,
            mlab=mlab, transf=transf, atransf=atransf, targs=targs,
            efac=efac, col=col, border=border, lty=lty, fonts=fonts, cex=cex, annosym=annosym, ...)