+++ R/predict.rma.ls.r
@@ -501,6 +501,7 @@ level, digits, transf, targs, vcov=FALSE, ...) {
    out$digits <- digits
    out$method <- x$method
    out$transf <- do.transf
+   out$pred.type <- ifelse(pred.mui, "location", "scale")
 
    if (x$test != "z")
       out$ddf <- ddf
@@ -508,7 +509,7 @@ level, digits, transf, targs, vcov=FALSE, ...) {
    if (pred.mui && (x$test != "z" || is.element(pi.type, c("riley","t"))) && pi.type != "simple")
       out$pi.ddf <- pi.ddf
 
-   class(out) <- "list.rma"
+   class(out) <- c("predict.rma", "list.rma")
 
    if (vcov & !do.transf) {
       out <- list(pred=out)